requirements.txt
azure-storage-blob==12.15.0
pandas==1.5.3
pip-chill==1.0.1
python-dotenv==1.0.0
